requirements.txt
werkzeug
Flask==1.1.1
pyzmq==18.1.1
Flask-JWT
Flask-SQLAlchemy